Split relational query binding
