+++ crates/truthdb-core/src/engine/relational/query/binding/mod.rs
@@ -1,0 +1,11 @@
+use super::super::prelude::*;
+
+mod functions;
+mod projection;
+mod scope;
+mod subquery;
+
+pub(in crate::engine::relational) use functions::*;
+pub(in crate::engine::relational) use projection::*;
+pub(in crate::engine::relational) use scope::*;
+pub(in crate::engine::relational) use subquery::*;
